+++ src/transaction/index.ts
@@ -1,0 +1,5 @@
+export { NoopTransactionPerformer } from './NoopTransactionPerformer.js';
+export type {
+	TransactionableAsync,
+	TransactionPerformer,
+} from './Transaction.js';